rename test files
also combine execute/subscribe
